refactor(layers): repoint tracing shim callers
