+++ aragora/events/async_dispatcher.py
@@ -25,6 +25,7 @@ import json
 import logging
 import os
 import time
+from collections.abc import Callable
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any
 
@@ -152,7 +153,7 @@ class AsyncWebhookDispatcher:
             Tuple of (success, status_code, error_message)
         """
         from aragora.security.webhook_signing import generate_signature
-        from aragora.server.middleware.tracing import get_trace_id
+        from aragora.observability.middleware.tracing import get_trace_id
 
         client = await self._ensure_client()
 
@@ -227,15 +228,25 @@ class AsyncWebhookDispatcher:
             AsyncDeliveryResult with outcome
         """
         # Import metrics and tracing lazily
+        record_webhook_retry: Callable[[str, int], None] | None
         try:
-            from aragora.observability.metrics.webhook import record_webhook_retry
+            from aragora.observability.metrics.webhook import (
+                record_webhook_retry as _record_webhook_retry,
+            )
         except ImportError:
             record_webhook_retry = None
+        else:
+            record_webhook_retry = _record_webhook_retry
 
+        trace_webhook_delivery: Callable[..., Any] | None
         try:
-            from aragora.observability.tracing import trace_webhook_delivery
+            from aragora.observability.tracing import (
+                trace_webhook_delivery as _trace_webhook_delivery,
+            )
         except ImportError:
             trace_webhook_delivery = None
+        else:
+            trace_webhook_delivery = _trace_webhook_delivery
 
         event_type = payload.get("event", "unknown")
         correlation_id = payload.get("correlation_id") or payload.get("data", {}).get(
